requirements.txt
requests
flask
pandas
python-dotenv
openai
python-telegram-bot
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gspread
